[GEOT-6419] gt-jdbc-hana - Missing apostrophe in query generation
If the SRID of a column is not known (e.g. the column is part of a
view), a malformed SQL query is generated. In particular, the geometry
string literal does not have the second enclosing apostrophe.
